utils: Move log_system_info() in here
